pyproject.toml
[tool.poetry]
name = "nextjs-fastapi"
version = "0.1.0"
description = ""
authors = ["raza panjwani"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
psycopg2 = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^2.6.4"}
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.28"
pytest = "^8.1.1"
bcrypt = "^4.1.2"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
